Initial Commit
This is a VERY BASIC game created for the course GTAT1 "Advanced Game
Programming" at my university. The task was to create a Shoot-Em-Up, so
I started making one. But then I read the requirements more closely,
such as a one-hit player death, infinitely respawning PASSIVE enemies
and other things, and had to adapt my current concept to these
constraints. I didn't put that much effort in keeping the parts created
for that task clean and I didn't bother removing all the stuff that's
not used here (Fonts etc).